+++ lib/FastOperation.dart
@@ -2,3 +2,5 @@ library FastOperation;
 export 'base_converter/base_converter.dart';
 export 'distance/distance.dart';
 export 'weight/weight.dart';
+export 'temperature/temperature.dart';
+export 'angle/angle.dart';